fix: refactor tracking for notifications
