+++ packages/shared/src/components/modals/SubmitArticleModal.tsx
@@ -23,7 +23,7 @@ import LinkIcon from '../icons/Link';
 import Alert, { AlertParagraph, AlertType } from '../widgets/Alert';
 import { Modal, ModalProps } from './common/Modal';
 import EnableNotification from '../notifications/EnableNotification';
-import { NotificationPromptSource } from '../../hooks/useEnableNotification';
+import { NotificationPromptSource } from '../../lib/analytics';
 
 type SubmitArticleModalProps = {
   headerCopy: string;
