+++ packages/shared/src/components/squads/SquadPageHeader.tsx
@@ -4,7 +4,6 @@ import { Squad, SourceMember, SourcePermissions } from '../../graphql/sources';
 import { SquadHeaderBar } from './SquadHeaderBar';
 import { SquadImage } from './SquadImage';
 import EnableNotification from '../notifications/EnableNotification';
-import { NotificationPromptSource } from '../../hooks/useEnableNotification';
 import { FlexCol } from '../utilities';
 import SquadMemberShortList from './SquadMemberShortList';
 import useSidebarRendered from '../../hooks/useSidebarRendered';
@@ -14,6 +13,7 @@ import TutorialGuide from '../tutorial/TutorialGuide';
 import { TourScreenIndex } from './SquadTour';
 import { useSquadTour } from '../../hooks/useSquadTour';
 import { verifyPermission } from '../../graphql/squads';
+import { NotificationPromptSource } from '../../lib/analytics';
 
 interface SquadPageHeaderProps {
   squad: Squad;
